Fix: Prevent delete() on null in AppSettingsFormTest tearDown

+++ tests/src/FunctionalJavascript/AppSettingsFormTest.php
@@ -48,12 +48,14 @@ class AppSettingsFormTest extends ApigeeEdgeFunctionalJavascriptTestBase {
   /**
    * {@inheritdoc}
    */
-  protected function tearDown(): void {
-    try {
-      $this->defaultApiProduct->delete();
-    }
-    catch (\Exception $exception) {
-      $this->logException($exception);
+protected function tearDown(): void {
+    if ($this->defaultApiProduct) {
+      try {
+        $this->defaultApiProduct->delete();
+      }
+      catch (\Exception $exception) {
+        $this->logException($exception);
+      }
     }
     parent::tearDown();
   }
